requirements.txt
numpy>=2.0.0
opencv-python>=4.8.0
pandas>=2.2.0
PyYAML>=6.0.0
pypdf>=4.0.0
